+++ src/main/java/ros/msgs/geometry_msgs/Vector3.java
